+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the "L13 Diff" extension will be documented in this file.
 ## [0.17.0] - 2019-07-07
 
 - Added ignore line endings for text files.
-- Added setting `l13Diff.eolFileExtenions` to extend the default whitelist for text files.
+- Added setting `l13Diff.ignoreEndOfLine`.
 - Added `L13 Diff` as first favorite to open diff panel with mouse.
 - Added dimmed dirnames in list view.
 - Added color for conflicts if file has been changed to folder or vice versa.